+++ Runtime/Extensions/MonoBehaviourExtension.cs
@@ -22,5 +22,16 @@ namespace DT.UniStart {
       }
       mb.StartCoroutine(InvokeRepeatingRoutine(f, delay, interval));
     }
+
+    /// <summary>
+    /// Try to get a component from the game object.
+    /// If it doesn't exist, add it to the game object and return it.
+    /// </summary>
+    public static T GetOrAddComponent<T>(this MonoBehaviour mb) where T : Component {
+      // IMPORTANT: don't use `??` to check for null, because Unity overrides the == operator
+      var res = mb.gameObject.GetComponent<T>();
+      if (res != null) return res;
+      return mb.gameObject.AddComponent<T>();
+    }
   }
 }